+++ tests/test_get.py
@@ -1,6 +1,6 @@
 import pytest
 
-from gazpacho.get import HTTPError, get, sanitize
+from gazpacho.get import HTTPError, get
 
 
 def test_get():
@@ -28,13 +28,3 @@ def test_HTTPError_404():
     url = "https://httpstat.us/404"
     with pytest.raises(HTTPError):
         get(url)
-
-
-def test_sanitize_weird_characters():
-    url = sanitize("https://httpbin.org/anything/drãke")
-    assert url == "https://httpbin.org/anything/dr%C3%A3ke"
-
-
-def test_sanitize_missing_protocol():
-    url = sanitize("gazpacho.xyz")
-    assert url == "http://gazpacho.xyz"
