+++ tests/test_utils.py
@@ -2,7 +2,7 @@ from xml.parsers.expat import ExpatError
 
 import pytest
 
-from gazpacho.utils import format, match, recover_html_and_attrs
+from gazpacho.utils import format, match, recover_html_and_attrs, sanitize
 
 
 def test_attr_match():
@@ -79,3 +79,13 @@ def test_format_fail():
     html = """<div><ul><li>Item</li><li>Item</li></ul>"""
     with pytest.raises(ExpatError):
         format(html, fail=True)
+
+
+def test_sanitize_weird_characters():
+    url = sanitize("https://httpbin.org/anything/drãke")
+    assert url == "https://httpbin.org/anything/dr%C3%A3ke"
+
+
+def test_sanitize_missing_protocol():
+    url = sanitize("gazpacho.xyz")
+    assert url == "http://gazpacho.xyz"
